Fix and enable crash report tests
This is a new attempt to solve the problem of the flaky tests.

Design Document:
https://docs.google.com/document/d/1_V0_W3aMxnD-dWYYSJkni8U5PGKZNoaB6SoE4N02TsA/edit?usp=sharing

Queued as task: A DoQueueReport task has been added to the backlog.

Fully added to cache: The report has been added to the reporting cache.

I believe there is a race condition in the Reporting API where reports
are dropped because their sources are marked as expired before the
reports are fully added to the cache.

1. A report is queued with reporting source.

2. SendReportsAndRemoveSource is called for this source.

3. No reports are found to deliver for this source (because the report
hasn't been fully added to cache yet).

4. The source is marked as expired.

5. Then when the report is actually attempted to be added to the cache
(ReportingCacheImpl::AddReport): It is dropped because its reporting
source is expired.

This CL addresses this by modifying SendReportsAndRemoveSource() to use
DoOrBacklogTask to queue as a task the expiration of the reporting
source itself. This ensures that the expiration task is placed at the
end of the backlog, after any previously queued tasks (such as adding
reports). QueueReport uses DoOrBackLogTask() to queue reports
(DoQueueReport), so by also using it when we expire reporting sources,
we maintain the ordering between calls to DoQueueReport and
ExpireReportingSource.

As a result, I believe we guarantee the correct ordering:

- Any report queued as a task before requesting source expiration will be processed first.
- Only after processing these reports will the source be marked as
expired.

You should also note that the backlog task queue is only ever used when
the ReportingService hasn't been fully initialized yet. This means that
this CL only fixes the race condition where:

ReportingService is not fully initialized.

QueueReport() comes in but doesn't finish—it gets queued as a backlog
task.

SendReportsAndRemoveSource() gets called because the renderer crashes
before ReportingService is initialized; the reporting source gets
expired immediately.

Eventually, ReportingService gets initialized, and the backlog task
queue is executed. The QueueReport() task is executed, but the report is
dropped because the reporting source has already been expired.

Prior attempt:
https://chromium-review.googlesource.com/c/chromium/src/+/5422052

Bug: 40268201
Change-Id: Id49f86a0b7ae4cf10b950792202d2f3819d70713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6406078
Feels: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Issack John <issackjohn@microsoft.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442519}