Remove use of `SkFontMgr::matchFamily()` in pdfium_font_win.cc
The rest of Chromium avoids `SkFontMgr::matchFamily()` for a few
reasons, mainly:
1. Skia devs have mentioned a desire to deprecate it
2. It's not particularly more useful than `matchFamilyStyle()` because
   code that wants to use a typeface to shape/draw glyphs will need to
   get it again with `matchFamilyStyle` and an appropriate `style`

In this instance, pdfium_font_win used `matchFamily()` only to check for
the existence of a family with a given name, then immediately called
`matchFamilyStyle()` with the family name and the known style. This is
problematic because:

1. There is no guarantee that a successful call to `matchFamily()` will
   yield a successful call to `matchFamilyStyle()` with any `style`
2. It's inefficient: `matchFamily()` and `matchFamilyStyle()` are each
   implemented as IPCs to the browser process
3. The DWriteFontProxy replacement (FontDataManager) purposely doesn't
   implement `matchFamily` for the 2 reasons listed in the first part
   of this CL message.

This CL changes the code that determines which family name to use with
code that instead directly invokes `matchFamilyStyle()` for the last
resort fallback typeface for asian languages in PDFs without embedded
fonts.

This shouldn't lead to any behavior change (% fixing the crash in the
linked bug), except in some rare cases where `matchFamily()` would have
succeeded but the following call to `matchFamilyStyle()` would've
failed. In those cases, we might try another fallback that could
potentially succeed.

Bug: 405534974
Change-Id: I5ac1382ee6fb68152f491dac569020baa20c9676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6420149
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442532}