[GTK] Divide text size by font scale on GTK4
This is similar to the cursor scaling issue [1] that was fixed recently.
On GTK4, the scale factor is included in the font size so must be
divided out.

This also includes a small fix for cursor scaling. Previously the
modified scale factor was used (which may be rounded, include the font
scale, and is clamped to >=1). This CL uses the unprocessed scale
instead. This fixes an issue where the font scale improperly affected
the cursor size.

Tested that UI scaling, font scaling, and cursor size is correct with
all combinations of:
{X11,Wayland}x{scale=1,scale=2}x{font_scale=1,font_scale=2}x{GTK3,GTK4}.

[1] https://crrev.com/1441877

R=thestig

Change-Id: Ieeae1f64c91ee1f042f7ec891a6a9e264ce38b40
Fixed: 407816395
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432677
Auto-Submit: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442549}