[views-ax] Fix Tooltip Text in `TRENDING SEARCHES` in omnibox.
The CL:6089787 introduces a regression where the tooltip text of the
`TRENDING SEARCHES` toggle button is incorrectly displayed when it
is hidden.
When the trending search results are hidden, the tooltip text
should be `Show Suggestions`, but instead, it shows
`Hide Suggestions`.

The primary goal of `CL:6089787` was to migrate tooltip text to the
new push mechanism, which it successfully achieves. However, an
issue arises when the button is toggled,
as the function `OmniboxHeaderView::UpdateUI()` is called
when the button is hovered over and then unhovered.
This function explicitly sets the TooltipText value using
`SetTooltipText()`. The main problem is that it always sets the
tooltip text to `Hide Suggestions`, even when the popup is hidden.
This discrepancy did not occur with the previous pull-based
mechanism but is now evident with the push-based mechanism.

To resolve this, the `UpdateTooltipText()` function of the toggle
button class is explicitly called within the
`OmniboxHeaderView::UpdateUI()` function. This ensures that the
correct tooltip text is displayed based on the toggle state
(i.e., whether the popup is open or not).

Fixed: 406455724
Change-Id: I243a5230fada3b4e86400ed2f00656b8a0aebb7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6424522
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442587}