Refactor GlicWindowController AuthCheckDone
The goal of this refactor is to make it more clear how execution flows
from the AuthCheckDone to finishing showing.

The primary refactor is to split OpenAttached and OpenDetached into a
part that sets up the glic widget and then a part that starts the
animation if needed.

Two minor refactors are:

* Rename: ShowFinish to GlicLoadedAndAnimationDone to better reflect
what is happening here.
* Add a kWaitingForAuthCheck state since it does not make sense to be
in the kOpenAnimation state while the auth check is being preformed.

There are no functional changes from this CL.

Bug: 391402352
Change-Id: Ic715c0c3163004e95a070de3c722a2246815c037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6423369
Reviewed-by: Dan Harrington <harringtond@chromium.org>
Commit-Queue: Trevor Perrier <perrier@chromium.org>
Auto-Submit: Trevor Perrier <perrier@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Reviewed-by: Sana Akbani <sanaakbani@google.com>
Cr-Commit-Position: refs/heads/main@{#1442603}