[Autofill] Optimize RandomizedEncoder a little
RandomizedEncoder includes two concepts: "encoding type" and
"encoding info".

The "type" is an integer that determines the "info".

Before this CL, the info was looked up by binary search in an array.
A pointer to that element was stored as a member of RandomizedEncoder.
The type was not explicitly stored, but still available as a member
of the info.

After this CL, RandomizedEncoder stores only the type, not the info
anymore. The info is accessed in an array, but rather than doing
binary search, we use the integer as index directly.

Bug: 40100455
Change-Id: I17f08670f9d0708f152daa4165ecc24602db646a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6430112
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1442623}