Undo crrev.com/c/3715413 introduced by crbug.com/237461655
This CL is pretty much an exact revert of crrev.com/c/3715413. Context:
crrev.com/c/371541 was introduced as a fix for crbug.com/237461655, and
was specifically intended for Lacros.

Now that Lacros is sunsetted, this CL hopes that by reverting
crrev.com/c/3715413, the simplicity of the code can be reduced back to
the pre-Lacros level.

Bug: b:354842935
Test: enabling this specific tast test http://cs/chromeos_public/src/platform/tast-tests/src/chromiumos/tast/local/bundles/cros/inputs/virtual_keyboard_typing_input_fields.go?rcl=91dce935c9edd08bf7261ef1503c146c642e6f2b&l=38
Change-Id: I159614925e0468d78f5e8b6d916fedbea8eadeeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6414330
Reviewed-by: John Palmer <jopalmer@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442684}