[gpu] Make GpuInfoCollector toggles in sync with DawnContextProvider
GpuInfoCollector::GetDawnTogglesForSkiaGraphite is currently out of
sync with the toggles enabled/disabled in DawnContextProvider.
Ideally the toggles in GpuInfoCollector should not be a separate list
and also come from DawnContextProvider. But for now, update
GpuInfoCollector to have most of the toggles similar to that in
DawnContextProvider along with TODOs for those that require extra
flags and are platform specific.

Bug: 407497928
Change-Id: Ib687a6f806ba9a846eeb5887bc16b836a373a932
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6416722
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442700}