Fixed the border state when switching to a null tab
When closing the current tab or creating a new tab using alt+enter, we
get a tab with null value in between the switching. This was previously
handled by introducing a new enum and state here: crrev.com/c/6254677.
This fix however, puts the border in a bad state when creating a new tab
using alt+enter.
Thus, in this change the `kFocusedTabDestroyed_NoFocusChange` state is
removed and instead we do the following:
- A -> null: start ramping down.
- null -> B: pick up the opacity and ramp up from its current value.

Credit: huge thanks to liuwilliam@ and vollick@ for the proposed solution.

Bug: 407106595
Change-Id: I5d35e1d30fe399d4afa3fc3b0c8b465ebb096a82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6419087
Commit-Queue: Baran Erfani <baranerf@google.com>
Reviewed-by: William Liu <liuwilliam@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442707}