Use SharedImage size instead of DrawingBuffer size
In few places we use size_ while interacting with SharedImage.
`size_` and `shared_image->size()` matches in this context because
shared images are creates with `size_` and recreated on the resize, but
semantically it's easier to reason about if it comes from the client
shared image, as it's the source of truth.

Change-Id: I1bedebcd7e413ddb94557d178ca34a3bdc9b794d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426306
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442754}