Implementation for send after new successful navigation feature
Introduces standby mode for the report scheduler timer, which halts all
report sends until a new successful navigation is received. It is
believed this will aid report delivery loss we're seeing, particularly
on the Android platform.

This feature is separate from the connection tracker, allowing us to
compare the two features and judge which will be more beneficial.

This feature is Finch flag-gated and will be tested on in an upcoming
experiment. The cutoff for what's considered to be in standby is
defaulted to [2 minutes] from the latest successful navigation, but this
value will also be experimented on.

Bug: 384870263
Change-Id: Idd7a4edbe92d2b9acea24ec360f0bad87f740e78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6106631
Commit-Queue: Thomas Quintanilla <tquintanilla@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442786}