Add trigger for tracing reports to CCT startup when warm
This is necessary for collecting chrometto histograms, and will
hopefully help explain why MobileStartup.IntentToCreationTime.CustomTabs
regresses as much as it does in the CCTEarlyNav experiment.

Also added a trace event to some code that I'm curious seeing
the performance of in the wild.

Bug: 372710946
Change-Id: I5cfdd098892a5eb7b4a2b1b54a78cd17a7497d26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434030
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442797}