Remove early abort for CPU raster mode in IsResourceValid
IsResourceValid has a confusing condition where it returns true if we
have a CanvasResourceProviderBitmap and the preferred_2d_raster_mode_ is
kPreferCPU.

The check on CanvasResourceProviderBitmap leverages the fact that once
created and valid, a CanvasResourceProviderBitmap can never become
invalid. Thus, we can skip the rest of the function. We could however
argue that this kind of logic breaks encapsulation: instead of making
assumptions, it would be better to call IsValid on the provider and let
it decide for itself it it's valid. This is what would happen without
the check.

The check on preferred_2d_raster_mode_ is far more confusing. We can
track its origin to https://codereview.chromium.org/2123623002,
submitted in 2016. The reason for the check isn't explicitly explained.
We can however observe that in this CL, once acceleration is disabled,
m_surface remains permanently locked on a non-accelerated surface. By
returning true if m_accelerationMode == DisableAcceleration in
Canvas2DLayerBridge::checkSurfaceValid() [1], the logic detecting a GPU
context loss and resetting the canvas was skipped, ensuring that the CPU
CPU surface isn't destroyed if there is a GPU context error.

This check likely became useless in 2017, with CL
https://crrev.com/c/758562, where the context loss logic was updated to
run only for accelerated CanvasResourceProvider. We've been maintaining
the check ever since. The code detecting GPU context losses and
resetting the canvas was later entirely removed from IsResourceValid,
in https://crrev.com/c/6387026. Thus, the early abort based on
preferred_2d_raster_mode_ is now useless.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_2d_layer_bridge.cc;l=749;drf=third_party%2FWebKit%2FSource%2Fplatform%2Fgraphics%2FCanvas2DLayerBridge.cpp;drc=627ddfd0de197d67567a15191f654b8835408bce

Bug: 346766781
Change-Id: I139410ff759c39e3f3842c4fef91c5dc96cb3725
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426233
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442862}