[UIA] Disable UIA provider when JAWS 2024 and older is running
We recently learned that JAWS 2024 and older do not work properly when
the UIA provider is enabled in Chromium. This is a JAWS issue, that they
resolved in older versions for Microsoft Edge (which supports the UIA
provider by default) and for Google Chrome as of JAWS 2025.

In order to rollout UIA to all Chrome users without creating issues for
the JAWS users, we are building this mechanism that detects when an
older version of JAWS is running and disables the UIA provider for those
users.

Disabling the UIA provider is two folds:
1. Set the UIA enablement to `kDisabled`, as it's done by the flag.
2. Notify UIA that the previous UIA providers we exposes aren't valid anymore. To do so, we use the UiaDisconnectAllProviders function [1].

Because of an issue with JAWS in local Chromium builds (just doesn't
work in text fields, irrespective of whether UIA is enabled or not, with
and without this change), we couldn't test the complete solution in the
local build and are moving to test this out in Chrome Canary, where JAWS
does work in text fields when the UIA provider is disabled. However, we
tested all individual parts (getting the JAWS version, disabling UIA
after a provider was shared with a client if JAWS 2024 or lower is
running, etc.) successfully. If this change does not fix the JAWS issue,
we intend to revert this change.

Furthermore, this change is gated behind a kill switch, allowing us to
disable it remotely if unexpected issues arise with JAWS or other
assistive technologies.

[1]:
https://learn.microsoft.com/en-us/windows/win32/api/uiautomationcoreapi/nf-uiautomationcoreapi-uiadisconnectallproviders

Bug: 406006402
Change-Id: Ia2bd65c2c2268c99f20cb7b323c1f842b1fdc009
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6423354
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442877}