[Mac] Switch order of rpath in helper apps in component builds
In component builds the helper apps link directly against the framework.
The framework is found via the runpath search path. The helper apps also
dlopen the framework within their main function by constructing a path
relative to the helper application.

The previous runpath search path ordering was resulting in the loader
finding the framework at the top level of the build directory while the
helper app dlopend the framework embedded within Chromium / Content
Shell. This caused the framework to be loaded twice. For Chromium this
was mostly harmless as all of its dependencies are linked into
libchrome_dll.dylib. For Content Shell this resulted in a lot of spew
from the Objective-C runtime about duplicate classes being initialized.

Switching the rpath order ensures that the framework is found within the
Chromium / Content Shell app, while other libraries continue to be found
within the top level of the build directory.

Change-Id: I33390cf9eca7b727b68af0255b5e76be7192acec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6348359
Commit-Queue: Mark Rowe <markrowe@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442896}