[iOS] Implement CreateUploadEventsRequest for ios_realtime_reporting_client.
This is the proto equivalent of the existing
base::Value::Dict IOSRealtimeReportingClient::GetContext(). There is an
ongoing effort of migrating from Dict used in context to proto for the
uploaded events. Once the migration is done, it will switch to use the
proto for the event reporting.

Bug: 394097677
Change-Id: I8958ac794a5b627d4c74acb5b6f7fcbbcf642cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6432146
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Commit-Queue: Xu Li <mxlg@google.com>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442916}