[TabGroupSyncService] Fix dangling ptr - TabGroupSyncDelegateBrowserTest
Prevents a dangling_ptr when a TabGroupSyncDelegateBrowserTest was being
destructed. The previous fix required each test to add themselves as an
observer of the TabGroupSyncService. This way the observation was
removed during test destruction.

This change moves this logic into SetUpOnMainThread() so that new tests
don't need to worry about cleaning up the observer on destruction.

Change-Id: Icca8cd32519a0ce01111b72fd011af08091e3727
Bug: 408039922
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434381
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442959}