[PWA] Make UninstallApps() API private, clean up fake version
This CL makes the following changes:
1. Makes the UninstallApps() API private and not be able to be overriden
as that is not a public interface of the ExternallyManagedAppManager.
2. Removes the handle_uninstall_request_callback_, and rewires tests to
use the public interface for testing instead, mimicking production more
closely.

The initial plan was also to do the same for InstallApps() in this same
CL, but a lot of existing tests use it wrongfully by completely
bypassing the install flow and causing problematic behavior. This is
punted to a follow-up.

While this is an overall no-op on production, this makes it easier to
remove FakeExternallyManagedAppManager in the future.

Bug: 407549320, 407797639
Change-Id: I76369de79d1479dbb683cc487b078c800ee26802
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6431760
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442980}