extensions: Port ComponentLoader to desktop Android
It is a dependency of ChromeExtensionRegistrarDelegate, which we are
trying to port to desktop Android.

Move one function body from ExtensionService to ExtensionRegistrar so
it can be accessed by ComponentLoader directly.

Convert ComponentLoader to a KeyedService so it can be accessed (in
future CLs) without accessing ExtensionsService.

Bug: 408454704

Change-Id: Ia5a4c88e4299d40ed07322a7c40c7be9e6af0a7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6431575
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443011}