extensions: Port ChromeExtensionRegistrarDelegate to desktop Android
And replace (most of) the stub delegate implementation we have been
using so far. This makes extension install and uninstall much closer
to Win/Mac/Linux behavior.

UnpackedInstaller is currently being worked on, so it is #ifdef'd out.
Likewise, theme support is #ifdef'd out as well. Bugs are filed for
both.

An ExtensionSpecialStoragePolicy instance is added to the profile for ENABLE_EXTENSIONS_CORE. A ManagementPolicy instance is added to
DesktopAndroidExtensionSystem. Both are dependencies of
ChromeExtensionRegistrarDelegate.

Bug: 394968360
Change-Id: I0891f420a01c23e90fff358ac6549b8589d0b002
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434815
Auto-Submit: James Cook <jamescook@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443016}