[PermissionOptions] Generalize PermissionRequestDescription
Refactor PermissionRequestDescription to hold PermissionDescriptorPtr.
Permissions with options can no longer be represented by a simple
PermissionType enum value, hence future infrastructure refactorings to
enable them will pass around PermissionDescriptorPtrs instead of mapping
PermissionDescriptorPts to PermissionTypes.

To reduce the complexity of this CL, it limits the scope of the
refactoring to PermissionRequestDescription, its creations and usages
and uses utility mapping functions to map
PermissionRequestDescriptionPtr usages of this object to the
PermissionTypes. I.e. it does not yet modify the infrastructure itself
yet to rely on PermissionRequestDescriptorPtr.

Since all ContentSetting permission types currently map to exactly one
PermissionDescriptorPtr and vice versa, the mapping for all content
setting permission types is unique and can rely on these mapping utility
functions. Permissions that want to make use of permission options in
the future, will create the correct PermissionRequestDescriptorPtr
instead of relying on the current mapping functions.

refactoring. We need to add mapping functions to reduce the scope of
this CL and avoid having to make all changes at once. These mapping
functions will not be necessary in fuchsia, as the infrastructure will
move to fully relying on a different type.

Bug: 405943540, 393053278
Fuchsia-Binary-Size: Increase is temporary. This CL is part of a large
Change-Id: I9d6e3e71385749787a973a61e41ce130665c860d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6387077
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Florian Jacky <fjacky@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1443356}